REFACTORED EVERYTHING RELATED TO BLOCKS
Blocks are now no longer stored as Boxes with dyn structs, but rather just IDs.
This *is* a downgrade from the original system, but it will allow me to finally
implement world saving. Additional data for blocks will be stored in special
containers.

This "upgrade" makes the debug build very slow, but the release is way faster.
This must be because of something with closures.